chore: extract suite orchestration into src/orchestrator.ts
Part of REAL-1037.
