+++ test/e2e/src/orchestrator.ts
@@ -1,0 +1,70 @@
+import { createClient, SupabaseClient } from "@supabase/supabase-js";
+import { ANON_KEY, SERVICE_KEY, dbPassword, DB_URL_ARG, env, PROJECT_URL, REALTIME_OPTS } from "./context.ts";
+import type { SuiteDescriptor } from "./runner.ts";
+import { log, printSummary, flushOtel, results, createSuiteTest } from "./runner.ts";
+import { stopClient } from "./helpers.ts";
+import { setup, cleanup } from "./fixtures.ts";
+
+// Lives outside runner.ts to avoid an import cycle: fixtures.ts (setup/cleanup) already
+// imports `log` from runner.ts, so runner.ts can't import fixtures.ts back.
+export async function runSuites(descriptors: SuiteDescriptor[], testCategories: string[] | null) {
+  const LOAD_SUITES = descriptors.map((d) => d.name).filter((n) => n.startsWith("load"));
+  const FUNCTIONAL_SUITES = descriptors.map((d) => d.name).filter((n) => !n.startsWith("load"));
+
+  const activeCategories = testCategories
+    ? testCategories.flatMap((c: string) => {
+        if (c === "functional") return FUNCTIONAL_SUITES;
+        if (c === "load") return LOAD_SUITES;
+        return [c];
+      })
+    : null;
+
+  if (activeCategories) {
+    const unknown = activeCategories.filter((c: string) => !descriptors.some((d) => d.name === c));
+    if (unknown.length > 0) {
+      const valid = ["functional", "load", ...descriptors.map((d) => d.name)].join(", ");
+      log(`Unknown test categories: ${unknown.join(", ")}\nValid categories: ${valid}`);
+      process.exit(1);
+    }
+  }
+
+  const suitesToRun = activeCategories
+    ? descriptors.filter((d) => activeCategories.includes(d.name))
+    : descriptors;
+
+  const needsDb = suitesToRun.some((d) => d.needsDb);
+
+  if (needsDb && !SERVICE_KEY) {
+    console.error("--secret-key is required");
+    process.exit(1);
+  }
+
+  if (needsDb && env !== "local" && !dbPassword && !DB_URL_ARG) {
+    console.error("--db-password is required for staging and prod environments");
+    process.exit(1);
+  }
+
+  let userId: string | null = null;
+  let testUser: { email: string; password: string } = { email: "", password: "" };
+  let supabase: SupabaseClient = createClient(PROJECT_URL, ANON_KEY, { realtime: REALTIME_OPTS });
+
+  if (needsDb) {
+    const setupResult = await setup();
+    userId = setupResult.userId;
+    testUser = setupResult.testUser;
+    supabase = setupResult.supabase;
+  }
+
+  const start = performance.now();
+  try {
+    for (const d of suitesToRun) await d.run({ testUser, supabase, test: createSuiteTest(d.label) });
+  } finally {
+    await stopClient(supabase);
+    if (userId) await cleanup(userId);
+  }
+
+  printSummary(performance.now() - start);
+  await flushOtel();
+
+  if (results.some((r) => !r.passed)) process.exit(1);
+}
